% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aml2_make_model.R
\name{aml2_make_model}
\alias{aml2_make_model}
\title{Function to train Deep Learning regression model}
\usage{
aml2_make_model(
  asset_dataset,
  asset_name,
  num_bars = 1500,
  timeframe = 60,
  path_data = path_data,
  path_model = path_model,
  setup_mode = FALSE
)
}
\arguments{
\item{asset_dataset}{Dataset containing price and corresponding indicator of the asset}

\item{asset_name}{String, name of the asset we work with}

\item{num_bars}{Number of bars used to train the model}

\item{timeframe}{Data timeframe e.g. 60 min}

\item{path_data}{Path where the data files are stored}

\item{path_model}{Path where the models are be stored}

\item{setup_mode}{When TRUE function will attempt to write model to the disk without checking it}
}
\value{
Function is output is a Deep Learning model object
}
\description{
Function is training h2o deep learning model to match future prices of the asset to the indicator pattern.
Main idea is to be able to predict future prices by relying on several indicators and their patterns
}
\details{
Several indicators data points on different timeframes are provided together with corresponding asset price.
Function is handling shift of the price and indicator datasets.
Function will also check how the model predict by using trading objective.
NOTE: This is a special variant of the function aml_make_model.R to handle data from files

Because of the function is intended to periodically re-train the model it would always check how the previous model was working.
In case new model is better, the better model will be used.
}
\examples{

\donttest{

library(dplyr)
library(readr)
library(magrittr)
library(h2o)
library(lazytrade)
# start h2o engine (using all CPU's by default)
h2o.init()

path_model <- normalizePath(tempdir(),winslash = "/")
path_data <- normalizePath(tempdir(),winslash = "/")

# using asset 1
asset_dataset = system.file("extdata", "AI_RSIADXUSDJPY.csv",
                            package = "lazytrade") \%>\% read_csv(col_names = F)

aml2_make_model(asset_dataset = asset_dataset,
                asset_name = "USDJPY",
                num_bars = 1500,
                timeframe = 60,
                path_data = path_data,
                path_model = path_model,
                setup_mode = FALSE)


# stop h2o engine
h2o.shutdown(prompt = FALSE)

#set delay to insure h2o unit closes properly before the next test
Sys.sleep(5)

}



}
\author{
(C) 2020 Vladimir Zhbanko
}
