% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aml_consolidate_results.R
\name{aml_consolidate_results}
\alias{aml_consolidate_results}
\title{Function to consolidate model test results for several models}
\usage{
aml_consolidate_results(
  timeframe = 15,
  used_symbols,
  path_model,
  path_sbxm,
  path_sbxs,
  min_quality = 0.75
)
}
\arguments{
\item{timeframe}{Integer, Data timeframe e.g. 60 min. This will be equal to 1 bar}

\item{used_symbols}{Vector, containing several financial instruments that were previously used
to test the model}

\item{path_model}{String, User path where the test results were stored}

\item{path_sbxm}{String, User path to the sandbox where file with strategy test results should be written (master terminal)}

\item{path_sbxs}{String, User path to the sandbox where file with strategy test results should be written (slave terminal)}

\item{min_quality}{Double, value typically from 0.25 to 0.9 to select the min threshold value}
}
\value{
Function is writing files into Decision Support System folders
}
\description{
Function is designed to evaluate test results of multiple models.
This is done to select only group of best models
}
\details{
Provide a modular facility to aggregate and update files
}
\examples{


library(dplyr)
library(magrittr)
library(readr)
library(lazytrade)

testpath <- normalizePath(tempdir(),winslash = "/")
path_model <- file.path(testpath, "_Model")
path_sbxm <- file.path(testpath, "_T1")
path_sbxs <- file.path(testpath, "_T3")
dir.create(path_model)
dir.create(path_sbxm)
dir.create(path_sbxs)

file.copy(from = system.file("extdata", "StrTest-EURGBPM15.csv", package = "lazytrade"),
          to = file.path(path_model, "StrTest-EURGBPM15.csv"), overwrite = TRUE)

file.copy(from = system.file("extdata", "StrTest-EURJPYM15.csv", package = "lazytrade"),
          to = file.path(path_model, "StrTest-EURJPYM15.csv"), overwrite = TRUE)

file.copy(from = system.file("extdata", "StrTest-EURNZDM15.csv", package = "lazytrade"),
          to = file.path(path_model, "StrTest-EURNZDM15.csv"), overwrite = TRUE)

file.copy(from = system.file("extdata", "StrTest-EURUSDM15.csv", package = "lazytrade"),
          to = file.path(path_model, "StrTest-EURUSDM15.csv"), overwrite = TRUE)

Pairs <- c("EURGBP","EURJPY","EURNZD","EURUSD")

aml_consolidate_results(timeframe = 15,
                        used_symbols = Pairs,
                        path_model = path_model,
                        path_sbxm = path_sbxm,
                        path_sbxs = path_sbxs,
                        min_quality = 0.75)




}
\author{
(C) 2021 Vladimir Zhbanko
}
