% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_optimize.R
\name{check_if_optimize}
\alias{check_if_optimize}
\title{Function check_if_optimize. Purpose of this function is to verify trading system functionality by analysing
profit factor on the last trades. Whenever trading robot has profit factor value below certain limit
function will write a file log indicating which trading systems need to be maintained.}
\usage{
check_if_optimize(x, path_trading_robot, num_trades_to_consider = 10,
  profit_factor_limit = 0.7, demo_mode = F)
}
\arguments{
\item{x}{- dataframe containing trading results}

\item{path_trading_robot}{- path of trading robot repository. must contain folder TEST with file Setup.csv.
File Setup.csv contains a table with magic numbers under test}

\item{num_trades_to_consider}{- Number of trades to calculate profit factor}

\item{profit_factor_limit}{- Limit below which trading robot considered not working properly}

\item{demo_mode}{- When true function uses data stored in the _TEST_DATA folder}
}
\value{
function is writing files
}
\description{
Whenever there will be not enough trades then empty file will be written to the destination
}
\examples{

\dontrun{



}

}
