% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data_trades}
\alias{data_trades}
\title{Table with Trade results samples}
\format{
A dataframe with several columns
\describe{
\item{MagicNumber}{Unique identifiers of the Trading Robots}
\item{TicketNumber}{Ticket Number of closed position}
\item{OrderStartTime}{Date and Time when order started}
\item{OrderCloseTime}{Date and Time when order closed}
\item{Profit}{Monetary result of the trade}
\item{Symbol}{Symbol of the Asset e.g. EURUSD}
\item{OrderType}{Order Type 0 - buy, 1 - sell}
}
}
\usage{
data_trades
}
\description{
Table with Trade results samples
}
\keyword{datasets}
