% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DFR}
\alias{DFR}
\title{Table with predicted price change}
\format{
A dataframe with one column
\describe{
\item{"MagicNumber.x}{Unique identifiers of the Trading Robots from Trade Log}
\item{TicketNumber}{Ticket Number of closed position}
\item{OrderStartTime}{Date and Time when order started}
\item{OrderCloseTime}{Date and Time when order closed}
\item{Profit}{Monetary result of the trade}
\item{Symbol}{Symbol of the Asset e.g. EURUSD}
\item{OrderType}{Order Type 0 - buy, 1 - sell}
\item{"CUMSUM_PNL}{Cumulative sum of Profit and Loss}

}
}
\usage{
DFR
}
\description{
Table with predicted price change
}
\keyword{datasets}
