% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlog.R
\name{dlog}
\alias{dlog}
\title{Create log difference distribution}
\usage{
dlog(x)
}
\arguments{
\item{x}{\itemize{
\item matrix with one or more column
}}
}
\value{
\itemize{
\item dataframe
}
}
\description{
Calculate log distribution and calculate difference within rows

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{

library(magrittr)
library(lazytrade)
m <- seq(1:1000) \%>\% as.matrix(10) \%>\% dlog()


}
