% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{EURUSDM15X75}
\alias{EURUSDM15X75}
\title{Table with indicator and price change dataset}
\format{
A dataframe with several columns
\describe{
\item{LABEL}{future price change}
\item{X1-X75}{Values of the macd indicator}
}
}
\usage{
EURUSDM15X75
}
\description{
Table with indicator and price change dataset
}
\keyword{datasets}
