% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_macroeconomic_event.R
\name{evaluate_macroeconomic_event}
\alias{evaluate_macroeconomic_event}
\title{Function used to evaluate market type situation by reading the file with Macroeconomic Events}
\usage{
evaluate_macroeconomic_event(setup_path, file_name, path_T1, path_T3)
}
\arguments{
\item{setup_path}{Path with trading robots magic numbers setup, typically file 'Setup.csv'}

\item{file_name}{String indicating name of the file where Macroeconomic event is written, typically
"01_MacroeconomicEvent.csv"}

\item{path_T1}{Path of the Terminal 1}

\item{path_T3}{Path of the Terminal 3}
}
\value{
Function will write files indicating to enable or disable trading systems to open new orders
}
\description{
Function is reading the content of the file 01_MacroeconomicEvent.csv.
Content of the file can be either 1 or 0.
1 - when Macro Economic event is present, 0 - when it's not.
Function will also read magic number of the trading robots. This is indicated in the file 'Setup.csv'.
Final outcome of the function is the series of files written to the destination directories.
These files will either enable or disable opening of new positions in the trading robots
#'
}
\details{
This function is used exclusively with Market Type recognition system.

Final evaluation will consist in writing a dedicated file with a simple information:

When Macro economic even is not present:

"Magic","IsEnabled"
8139125,1

or, when Macro economic event is present:

"Magic","IsEnabled"
8139125,0
}
\examples{

# evaluate data on macroeconomic event (required to start trading)
library(tidyverse)

dir <- tempfile()

evaluate_macroeconomic_event(setup_path = system.file("extdata",
                                                 "Setup.csv",
                                                 package = "lazytrade"),
                             file_name = system.file("extdata",
                                                 "01_MacroeconomicEvent.csv",
                                                 package = "lazytrade"),
                             path_T1 = "", path_T3 = dir)

#unlink("/tmp/*.csv", recursive = TRUE, force = TRUE)

}
