% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_RL_policy_mt.R
\name{generate_RL_policy_mt}
\alias{generate_RL_policy_mt}
\title{Function performs RL and generates model policy for each Market Type}
\usage{
generate_RL_policy_mt(x, states, actions, control)
}
\arguments{
\item{x}{- Dataframe containing trading data}

\item{states}{- possible states for Reinforcement Learning}

\item{actions}{- possible actions}

\item{control}{- control parameters}
}
\value{
Function returns data frame with reinforcement learning model policy
}
\description{
This function will perform Reinforcement Learning using Trading Data.
It will suggest whether or not it is better to keep using trading systems or not.
Function is just using results of the past perfrormance to generate the recommendation (not a holy grail).
}
\details{
Initial policy is generated using a dummy zero values.
This way function starts working directly from the first observation.
However policy 'ON' value will only be generated once the Q value is greater than zero
}
\examples{

\dontrun{

generate_RL_policy_mt(data_trades, states, actions, control)

}

}
