% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{indicator_dataset}
\alias{indicator_dataset}
\title{Table with indicator dataset}
\format{A dataframe with several columns
\describe{
  \item{X1}{Date and time of the indicator sample}
  \item{X2-X29}{Values of the assets}
}}
\usage{
indicator_dataset
}
\description{
Table with indicator dataset
}
\keyword{datasets}
