% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{indicator_dataset_big}
\alias{indicator_dataset_big}
\title{Table with indicator dataset, 30000 rows}
\format{
A dataframe with several columns
\describe{
  \item{X1}{Date and time of the indicator sample}
  \item{X2-X29}{Values of the assets}
}
}
\usage{
indicator_dataset_big
}
\description{
Table with indicator dataset, 30000 rows
}
\keyword{datasets}
