% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{macd_100}
\alias{macd_100}
\title{Table with indicator only used to train model, 128 rows}
\format{
A dataframe with several columns
\describe{
  \item{EURUSD}{Values of the macd indicator}
  \item{GBPUSD}{Values of the macd indicator}
  \item{AUDUSD}{Values of the macd indicator}
  \item{NZDUSD}{Values of the macd indicator}
  \item{USDCAD}{Values of the macd indicator}
  \item{USDCHF}{Values of the macd indicator}
  \item{USDJPY}{Values of the macd indicator}
  \item{EURGBP}{Values of the macd indicator}
  \item{EURJPY}{Values of the macd indicator}
  \item{EURCHF}{Values of the macd indicator}
  \item{EURNZD}{Values of the macd indicator}
  \item{EURCAD}{Values of the macd indicator}
  \item{EURAUD}{Values of the macd indicator}
  \item{GBPAUD}{Values of the macd indicator}
  \item{GBPCAD}{Values of the macd indicator}
  \item{GBPCHF}{Values of the macd indicator}
  \item{GBPJPY}{Values of the macd indicator}
  \item{GBPNZD}{Values of the macd indicator}
  \item{AUDCAD}{Values of the macd indicator}
  \item{AUDCHF}{Values of the macd indicator}
  \item{AUDJPY}{Values of the macd indicator}
  \item{AUDNZD}{Values of the macd indicator}
  \item{CADJPY}{Values of the macd indicator}
  \item{CHFJPY}{Values of the macd indicator}
  \item{NZDJPY}{Values of the macd indicator}
  \item{NZDCAD}{Values of the macd indicator}
  \item{NZDCHF}{Values of the macd indicator}
  \item{CADCHF}{Values of the macd indicator}

}
}
\usage{
macd_100
}
\description{
Table with indicator only used to train model, 128 rows
}
\keyword{datasets}
