% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{macd_df}
\alias{macd_df}
\title{Table with one column indicator dataset}
\format{
A dataframe with one column
\describe{
\item{CADCHF}{Indicator values of the asset}
}
}
\usage{
macd_df
}
\description{
Table with one column indicator dataset
}
\keyword{datasets}
