% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{macd_ML2_small}
\alias{macd_ML2_small}
\title{Table with indicator and market type category}
\format{
A dataframe with several columns
\describe{
  \item{X1-X128}{Values of the macd indicator}
  \item{X129}{Category of Market Type}
}
}
\usage{
macd_ML2_small
}
\description{
Table with indicator and market type category
}
\keyword{datasets}
