% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{macd_ML60M}
\alias{macd_ML60M}
\title{Table with indicator and market type category used to train model}
\format{
A dataframe with several columns
\describe{
  \item{X1-X64}{Values of the macd indicator}
  \item{M_T}{Category of Market Type}
}
}
\usage{
macd_ML60M
}
\description{
Table with indicator and market type category used to train model
}
\keyword{datasets}
