% Generated by roxygen2: do not edit by hand
% Please edit documentation in www.vantharp.com/>
\name{mt_use_stats}
\alias{mt_use_stats}
\title{Function to use statistical method for Market Type recognition}
\usage{
mt_use_stats(
  indicator_dataset,
  symbol,
  num_bars = 64,
  timeframe = 60,
  path_data,
  path_sbxm,
  path_sbxs
)
}
\arguments{
\item{indicator_dataset}{Dataframe, One column containing indicator patterns to train the model}

\item{symbol}{Character symbol of the asset}

\item{num_bars}{Integer, Number of bars used to detect pattern}

\item{timeframe}{Integer, Data timeframe in Minutes.}

\item{path_data}{String, User path where the latest price data is stored}

\item{path_sbxm}{String, User path to the sandbox where file with strategy test results should be written (master terminal)}

\item{path_sbxs}{String, User path to the sandbox where file with strategy test results should be written (slave terminal)}
}
\value{
Log returns dataframe
}
\description{
Function is using log return distribution of financial asset's price.
Main idea is to estimate Market Type by solely relying on the current price pattern.
Latest price is obtained from the file, typically AI_CP60-300.csv. Log returns are
being calculated for each column

Selected Market Periods according to the theory from Van K. Tharp:
1. Bull normal, BUN
2. Bull volatile, BUV
3. Bear normal, BEN
4. Bear volatile, BEV
5. Sideways quiet, RAN
6. Sideways volatile, RAV
}
\details{
Function is using arbitrary defined levels of distributions.
}
\examples{

\donttest{

library(dplyr)
library(magrittr)
library(readr)
library(lazytrade)
library(ggplot2)
library(lubridate)

path_data <- normalizePath(tempdir(),winslash = "/")
path_sbxm <- normalizePath(tempdir(),winslash = "/")
path_sbxs <- normalizePath(tempdir(),winslash = "/")

ind = system.file("extdata", "AI_CP60-300.csv",
                  package = "lazytrade") \%>\% read_csv(col_names = FALSE)

ind$X1 <- ymd_hms(ind$X1)
tick = system.file("extdata", "TickSize_AI_RSIADX.csv",
                  package = "lazytrade") \%>\% read_csv(col_names = FALSE)

write_csv(ind, file.path(path_data, "AI_CP60-300.csv"), col_names = FALSE)
write_csv(tick, file.path(path_data, "TickSize_AI_RSIADX.csv"), col_names = FALSE)

ind1 <- select(ind, X2)
#'
# performing Deep Learning Regression using the custom function
mt_use_stats(indicator_dataset = ind1,
             symbol = 'EURUSD',
             num_bars = 64,
             timeframe = 60,
             path_data = path_data,
             path_sbxm = path_sbxm,
             path_sbxs = path_sbxs)


#set delay to insure h2o unit closes properly before the next test
Sys.sleep(5)

}



}
\author{
(C) 2020 Vladimir Zhbanko
}
