% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{policy_tr_systDF}
\alias{policy_tr_systDF}
\title{Table with Market Types and sample of actual policy for those states}
\format{
A dataframe with 2 columns:
\describe{
\item{MarketType}{Current Market Type status}
\item{Policy}{Policy choice}
}
}
\usage{
policy_tr_systDF
}
\description{
Table with Market Types and sample of actual policy for those states
}
\keyword{datasets}
