% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{price_dataset}
\alias{price_dataset}
\title{Table with price dataset}
\format{
A dataframe with several columns
\describe{
  \item{X1}{Date and time of the price sample}
  \item{X2-X29}{Values of the assets}
}
}
\usage{
price_dataset
}
\description{
Table with price dataset
}
\keyword{datasets}
