% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profit_factor.R
\name{profit_factor}
\alias{profit_factor}
\title{Calculate Profit Factor}
\usage{
# sumarizing table
 DF_Stats <- DF_Stats %>%
    filter(X3 > as.POSIXct(input$filterDate)) %>%
    group_by(X1) %>%
    summarise(PnL = sum(X5),
              NumTrades = n(),
              PrFact = profit_factor(X5))
}
\arguments{
\item{x}{column vector with profit or loss of the orders for one system}
}
\value{
function should calculate profit factor for this vector and return one value also as vector
}
\description{
Function is capable to calculate profit factor using a vector of the trading results
}
\examples{

# require(magrittr)
x <- read_rds('test_data.rds') \%>\% select(X5) \%$\% X5
x <- read_rds('test_data.rds') \%>\% filter(X5 > 0) \%>\% select(X5) \%$\% X5
x <- read_rds('test_data.rds') \%>\% filter(X5 < 0) \%>\% select(X5) \%$\% X5

}
