% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{profit_factor_data}
\alias{profit_factor_data}
\title{Table with Trade results samples}
\format{
A dataframe with several columns
\describe{
  \item{X1}{Unique identifiers of the Trading Robots}
  \item{X2}{Ticket Number of closed position}
  \item{X3}{Date and Time when order started}
  \item{X4}{Date and Time when order closed}
  \item{X5}{Monetary result of the trade}
  \item{X6}{Symbol of the Asset e.g. EURUSD}
  \item{X7}{Order Type 0 - buy, 1 - sell}
}
}
\usage{
profit_factor_data
}
\description{
Table with Trade results samples
}
\keyword{datasets}
