% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_policy.R
\name{record_policy}
\alias{record_policy}
\title{Record Reinforcement Learning Policy.}
\usage{
record_policy(x, last_result, trading_system, path_terminal)
}
\arguments{
\item{x}{- Dataframe containing columns MarketType and Policy}

\item{last_result}{- character vector of the last result of the trade}

\item{trading_system}{- character vector of lenght 1 with Trading System Magic Number information}

\item{path_terminal}{- path to the sandbox where this Policy/Decision must be written}
}
\value{
function creates csv file
}
\description{
Function will write a policy 'decision' to the csv file specific for each Expert Advisor
}
\examples{

\dontrun{

data(TradeStatePolicy)
trading_system <- 8118101
path_terminal <- "C:/Program Files (x86)/FxPro - Terminal3/MQL4/Files/"
record_policy(x = policy_tr_systDF, trading_system = trading_system, path_terminal = path_T4)

}


}
