% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_policy_mt.R
\name{record_policy_mt}
\alias{record_policy_mt}
\title{Record Reinforcement Learning Policy for Market Types}
\usage{
record_policy_mt(x, trading_system, path_sandbox)
}
\arguments{
\item{x}{- Dataframe containing columns MarketType and Policy}

\item{trading_system}{- character vector of lenght 1 with Trading System Magic Number information}

\item{path_sandbox}{- path to the sandbox where this Policy/Decision must be written}
}
\value{
nothing is returned but function will write csv file
}
\description{
Function will write a policy 'decision' to the csv file specific for each Expert Advisor
}
\examples{


library(tidyverse)
library(magrittr)
data(policy_tr_systDF)

record_policy_mt(x = policy_tr_systDF,
                 trading_system = 8118101,
                 path_sandbox = tempfile())


}
