% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{result_prev}
\alias{result_prev}
\title{Table with one column as result from the model prediction}
\format{
A dataframe with one column
\describe{
\item{predict}{Predicted values from the model}
}
}
\usage{
result_prev
}
\description{
Table with one column as result from the model prediction
}
\keyword{datasets}
