% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{result_R1}
\alias{result_R1}
\title{Table with aggregated trade results}
\format{
A dataframe with one column
\describe{
\item{predict}{predicted price change}
}
}
\usage{
result_R1
}
\description{
Table with aggregated trade results
}
\keyword{datasets}
