% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_write_control_parameters_mt.R
\name{rl_write_control_parameters_mt}
\alias{rl_write_control_parameters_mt}
\title{Function to find and write the best control parameters.}
\usage{
rl_write_control_parameters_mt(
  x,
  path_control_files,
  num_trades_to_consider = 100
)
}
\arguments{
\item{x}{\itemize{
\item dataset containing the trading results for one trading robot
}}

\item{path_control_files}{\itemize{
\item path where control parameters will be saved
}}

\item{num_trades_to_consider}{\itemize{
\item number of last trades to use for RL modeling simulations, default value 100
}}
}
\value{
Function writes best control parameters to be used by the Reinforcement Learning Function
}
\description{
This function is supposed to run on a weekly basis. Purpose of this function is to perform RL and
trading simulation and find out the best possible control parameters for the RL function.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Function is used by the R script Adapt_RL_MT_control.R
}
\examples{

\donttest{
# test lasts 15 sec:
dir <- normalizePath(tempdir(),winslash = "/")

library(dplyr)
library(readr)
library(ReinforcementLearning)
library(magrittr)
library(lazytrade)
data(trading_systemDF)

# use optimal control parameters found by auxiliary function
rl_write_control_parameters_mt(x = trading_systemDF,
                               path_control_files = dir,
                               num_trades_to_consider = 100)
}


}
\author{
(C) 2019, 2021 Vladimir Zhbanko
}
