% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{test_data_pattern}
\alias{test_data_pattern}
\title{Table with several columns containing indicator values and Label values}
\format{A dataframe with several columns
\describe{
  \item{LABEL}{Asset values as were recorded in the future}
  \item{V1-V49}{Transposed values of the indicator}
}}
\usage{
test_data_pattern
}
\description{
Table with several columns containing indicator values and Label values
}
\keyword{datasets}
