% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_m.R
\name{to_m}
\alias{to_m}
\title{Function converting time series data to matrix}
\usage{
to_m(x, n_cols)
}
\arguments{
\item{x}{- dataframe with one column}

\item{n_cols}{- number of columns in the matrix}
}
\value{
- matrix with specified amount of rows
}
\description{
PURPOSE: Transform Time Series Column of the dataframe to the matrix with specified number of columns.
Number of rows will be automatically found. Eventually remaining data points of the last row will be discarded
}
\examples{

\dontrun{

macd_m <- macd_df \%>\% select(X5.x) \%>\% to_m(64)

}



}
