% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TradeStatePolicy}
\alias{TradeStatePolicy}
\title{Table with Trade States and sample of actual policy for those states}
\format{A dataframe with 2 columns:
\describe{
  \item{TradeState}{Current trade state status}
  \item{Policy}{Policy choice}
}}
\usage{
TradeStatePolicy
}
\description{
Table with Trade States and sample of actual policy for those states
}
\keyword{datasets}
