% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{trading_systemDF}
\alias{trading_systemDF}
\title{Table with trade data and joined market type info}
\format{
A dataframe with several columns
\describe{
  \item{"MagicNumber.x}{Unique identifiers of the Trading Robots from Trade Log}
  \item{TicketNumber}{Ticket Number of closed position}
  \item{OrderStartTime}{Date and Time when order started}
  \item{OrderCloseTime}{Date and Time when order closed}
  \item{Profit}{Monetary result of the trade}
  \item{Symbol}{Symbol of the Asset e.g. EURUSD}
  \item{OrderType}{Order Type 0 - buy, 1 - sell}
  \item{"MagicNumber.y}{Unique identifiers of the Trading Robots from Ticket Opening Log}
  \item{"MarketType}{Logged Market Type of the asset at the moment of Ticket Opening}
}
}
\usage{
trading_systemDF
}
\description{
Table with trade data and joined market type info
}
\keyword{datasets}
