% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_control_parameters.R
\name{write_control_parameters}
\alias{write_control_parameters}
\title{Function to find and write the best control parameters.}
\usage{
write_control_parameters(x, path_control_files)
}
\arguments{
\item{x}{- dataset containing the trading results for one trading robot}

\item{path_control_files}{- path where control parameters will be saved}

\item{record_actual}{- if TRUE function will record actual results to the rds file;
if FALSE function will directly return best control parameters (very slow execution)}
}
\value{
Function writes best control parameters to be used by the Reinforcement Learning Function
}
\description{
This function is supposed to run on a weekly basis. Purpose of this function is to perform RL and
trading simulation and find out the best possible control parameters for the RL function.
}
\details{
Function is used by the R script Adapt_RL_control.R
}
\examples{

\dontrun{

write_control_parameters(trading_systemDF, path_control_files = path_control_files)

}

}
\author{
(C) 2019 Vladimir Zhbanko
}
