% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{x_test_model}
\alias{x_test_model}
\title{Table with a dataset to test the Model}
\format{
A dataframe with several columns
\describe{
  \item{X1}{future price change}
  \item{X2-X76}{Values of the macd indicator}
}
}
\usage{
x_test_model
}
\description{
Table with a dataset to test the Model
}
\keyword{datasets}
