% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{y}
\alias{y}
\title{Table with indicators and price change which is used to train model}
\format{
A dataframe with several columns
\describe{
\item{X1}{Time index}
\item{X2}{Closed price now}
\item{X3}{Closed price 34 bars ago}
\item{X4-X19}{Series of Indicator values}
\item{LABEL}{Price difference, difference between X3 and X2}
}
}
\usage{
y
}
\description{
Table with indicators and price change which is used to train model
}
\keyword{datasets}
