% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aml_collect_data.R
\name{aml_collect_data}
\alias{aml_collect_data}
\title{Function to read new data, transform data, save data for further retraining of regression model for a single currency pair}
\usage{
aml_collect_data(indicator_dataset, symbol, timeframe, path_data)
}
\arguments{
\item{indicator_dataset}{Dataset containing assets indicator which pattern will be used as predictor}

\item{symbol}{Character symbol of the asset for which to train the model}

\item{timeframe}{Data timeframe e.g. 1 min}

\item{path_data}{Path where the aggregated historical data is stored, if exists in rds format}
}
\value{
Function is writing files into Decision Support System folder, mainly file object with the model
}
\description{
Function is collecting data from the files
Data object are transformed to be suitable for Regression Modelling.
Price change will be in the column 'LABEL', column X1 will keep the time index
Result would be written to a new or aggregated to the existing '.rds' file

Function is also checking that generated dataset is not too big.
Should the dataset is too big (e.g. > 10000 rows), then only latest 9500 rows will be used.
Note: the amount 10000 rows is not verified in practice, further testing is required.
}
\details{
Function is handling shift of the price and indicator datasets.

This function is relying on the data collection from the dedicated data robot
Other 'aml_*' functions will work based on the data processed by this function
}
\examples{

# write examples for the function
library(dplyr)
library(readr)
library(lubridate)
library(lazytrade)


ind = system.file("extdata", "AI_RSIADXUSDJPY60.csv",
                  package = "lazytrade") \%>\% read_csv(col_names = F)

ind$X1 <- ymd_hms(ind$X1)

path_data <- normalizePath(tempdir(),winslash = "/")


# data transformation using the custom function for one symbol
aml_collect_data(indicator_dataset = ind,
                 symbol = 'USDJPY',
                 timeframe = 60,
                 path_data = path_data)


}
\author{
(C) 2020 Vladimir Zhbanko
}
