% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aml_score_data.R
\name{aml_score_data}
\alias{aml_score_data}
\title{Function to score new data and predict change for each single currency pair}
\usage{
aml_score_data(symbol, num_bars, timeframe, path_model, path_data,
  path_sbxm, path_sbxs)
}
\arguments{
\item{symbol}{Character symbol of the asset for which the model shall predict}

\item{num_bars}{Number of bars used to detect pattern}

\item{timeframe}{Data timeframe e.g. 1 min}

\item{path_model}{Path where the models are be stored}

\item{path_data}{Path where the aggregated historical data is stored, if exists in rds format}

\item{path_sbxm}{Path to the sandbox where file with predicted price should be written (master terminal)}

\item{path_sbxs}{Path to the sandbox where file with predicted price should be written (slave terminal)}
}
\value{
Function is writing file into Decision Support System folder, mainly file with price change prediction in pips
}
\description{
Function is using the latest data from the financial assets indicator pattern and deep learning model.
Prediction is a price change in the future for that asset will be used by the trading system
}
\details{
Performs fresh data reading from the rds file
}
\examples{

\donttest{

# test of function aml_make_model is duplicated here
library(readr)
library(h2o)
library(lazytrade)

path_model <- normalizePath(tempdir(),winslash = "/")
path_data <- normalizePath(tempdir(),winslash = "/")

data(EURUSDM15X75)
write_rds(EURUSDM15X75, file.path(path_data, 'EURUSDM15X75.rds'))

# start h2o engine (using all CPU's by default)
h2o.init()


# performing Deep Learning Regression using the custom function
aml_make_model(symbol = 'EURUSD',
               num_bars = 75,
               timeframe = 15,
               path_model = path_model,
               path_data = path_data)


path_sbxm <- normalizePath(tempdir(),winslash = "/")
path_sbxs <- normalizePath(tempdir(),winslash = "/")


# score the latest data to generate predictions for one currency pair
aml_score_data(symbol = 'EURUSD',
               num_bars = 75,
               timeframe = 15,
               path_model = path_model,
               path_data = path_data,
               path_sbxm = path_sbxm,
               path_sbxs = path_sbxs)

# stop h2o engine
h2o.shutdown(prompt = F)

}



}
\author{
(C) 2019 Vladimir Zhbanko
}
