% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_mykeys.R
\name{decrypt_mykeys}
\alias{decrypt_mykeys}
\title{Function that decrypt encrypted content}
\usage{
decrypt_mykeys(path_encrypted_content, path_private_key)
}
\arguments{
\item{path_encrypted_content}{- path to the encrypted content of the API key}

\item{path_private_key}{- path to the private RSA key, should be without password}
}
\value{
- a string with decrypted key
}
\description{
Function that decrypt encrypted content
}
\details{
It is possible to generate private/public key pair using R-Studio Project Options Menu.
Alternatively possible to use 'openssl' R package
}
\examples{

\dontrun{

# Consumer API keys
ConsumerAPIkeys <- decrypt_mykeys(path_encrypted_content = file.path(path_encrypted_keys,
                                  "ConsumerAPIkeys.enc.rds"),
                                  path_private_key = path_private_key)

}


}
