% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlog.R
\name{dlog}
\alias{dlog}
\title{Create log difference distribution}
\usage{
dlog(x)
}
\arguments{
\item{x}{- matrix with one or more column}
}
\value{
- dataframe
}
\description{
Transforms dataframe to the matrix
}
\examples{

library(magrittr)
library(lazytrade)
m <- seq(1:1000) \%>\% as.matrix(10) \%>\% dlog()


}
