% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profit_factorDF.R
\name{get_profit_factorDF}
\alias{get_profit_factorDF}
\title{Function that returns the profit factors of the systems in a form of a DataFrame}
\usage{
get_profit_factorDF(x, num_orders)
}
\arguments{
\item{x}{- data frame with orders. Note x must contain MagicNumber and Profit columns!}

\item{num_orders}{- desired number of orders to base profit factor calculation}
}
\value{
- Function returns dataframe with column PrFact with calculated profit factor value for each trading robot
}
\description{
Function that returns the profit factors of the systems in a form of a DataFrame
}
\examples{


library(dplyr)
data(profit_factorDF)
get_profit_factorDF(profit_factorDF, 10)



}
