% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_asset_data.R
\name{load_asset_data}
\alias{load_asset_data}
\title{Load Asset Data Function}
\usage{
load_asset_data(path_terminal, trade_log_file, time_period = 1,
  data_deepth = 50000)
}
\arguments{
\item{path_terminal}{- path to the MT4 terminal, string}

\item{trade_log_file}{- csv file name where the data is stored, without ".csv"}

\item{time_period}{- data periodicity in minutes, can be 1, 15, 60}

\item{data_deepth}{- collected data deepth in rows. describe how many rows in original file to read}
}
\value{
- dataframe with asset data in columns where X1 column is in a POSIXct format
}
\description{
Function imports file and changes date column type. Function return the dataframe with trade data.
-> update: data_deepth argument was added to optimize code and separate data needed to train and score the model
}
\details{
https://www.udemy.com/self-learning-trading-robot/?couponCode=LAZYTRADE7-10
}
\examples{

\dontrun{

# path_terminal <- "C:/Program Files (x86)/FxPro - Terminal2/MQL4/Files"
# path_terminal <- file.path(getwd(), "test_data")
# trade_log_file <- "AI_CP"
# trade_log_file <- "AI_Macd"
# time_period <- 1
# data_deepth <- "50000"
library(tidyverse)
library(lubridate)
prices <- load_asset_data(path_terminal = "C:/Program Files (x86)/FxPro - Terminal2/MQL4/Files",
                    trade_log_file = "AI_CP",
                    time_period = 1,
                    data_deepth = "50000")

}

}
