% Generated by roxygen2: do not edit by hand
% Please edit documentation in www.vantharp.com/>
\name{mt_stat_transf}
\alias{mt_stat_transf}
\title{Perform Statistical transformation and clustering of Market Types on the price data}
\usage{
mt_stat_transf(
  indicator_dataset,
  num_bars = 64,
  timeframe = 60,
  path_data,
  mt_classes
)
}
\arguments{
\item{indicator_dataset}{Dataframe, multiple column dataset containing price data in each column.
Each row is a time index, multiple columns are required but not strictly needed}

\item{num_bars}{Integer, Number of bars used to perform transformation}

\item{timeframe}{Integer, Data timeframe in Minutes, only used for naming convention}

\item{path_data}{String, User path where the dataset could be stored for the future use by other function}

\item{mt_classes}{Character Vector, with 2 or more Market Type classes}
}
\value{
Dataframe with statistically transformed and classified dataset for classification modeling
}
\description{
Function to statistically transform and cluster price data.
Multiple statistical properties are calculated for a defined time interval.
Once combined unsupervised learning is performed to assign 6 classes
}
\details{
Function will randomly assign Market Period labels based on Unsupervised Learning.
}
\examples{


library(dplyr)
library(stats)
library(magrittr)
library(readr)
library(lazytrade)
library(moments)

path_data <- normalizePath(tempdir(),winslash = "/")

data(price_dataset_big)

mt_stat_transf(indicator_dataset = price_dataset_big,
               num_bars = 64,
               timeframe = 60,
               path_data = path_data,
               mt_classes = c('BUN', 'BEN', 'RAN'))



}
\author{
(C) 2021 Vladimir Zhbanko
}
