% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profit_factor.R
\name{profit_factor}
\alias{profit_factor}
\title{PROFIT FACTOR FUNCTION}
\usage{
profit_factor(x)
}
\arguments{
\item{x}{column vector with profit or loss of the orders for one system}
}
\value{
function should calculate profit factor for this vector and return one value also as vector
}
\description{
PROFIT FACTOR FUNCTION
}
\examples{

# require(magrittr)
x <- read_rds('test_data.rds') \%>\% select(X5) \%$\% X5
x <- read_rds('test_data.rds') \%>\% filter(X5 > 0) \%>\% select(X5) \%$\% X5
x <- read_rds('test_data.rds') \%>\% filter(X5 < 0) \%>\% select(X5) \%$\% X5

}
