% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{result_R}
\alias{result_R}
\title{Table with predicte price change}
\format{
A dataframe with one column
\describe{
\item{predict}{predicted future price change}

}
}
\usage{
result_R
}
\description{
Table with predicte price change
}
\keyword{datasets}
