% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_m.R
\name{to_m}
\alias{to_m}
\title{Convert time series data to matrix with defined number of columns}
\usage{
to_m(x, n_cols)
}
\arguments{
\item{x}{- dataframe with one column}

\item{n_cols}{- number of columns in the matrix}
}
\value{
- matrix with specified amount of rows
}
\description{
Transforms Time Series Column of the dataframe to the matrix with specified number of columns.
Number of rows will be automatically found. Eventually not complete last row will be discarded
}
\examples{

library(magrittr)
macd_m <- seq(1:1000) \%>\% as.data.frame() \%>\% to_m(64)


}
